refactor(fpss): extract io_loop.rs
Move the I/O worker (io_loop), the login handshake (wait_for_login,
LoginResult), the is_read_timeout helper, the MAX_RECONNECT_ATTEMPTS
constant, and the ping heartbeat (ping_loop) out of fpss/mod.rs into a
dedicated io_loop submodule. The MAX_RECONNECT_ATTEMPTS smoke test moves
alongside the constant. Public items exposed to mod.rs are pub(super); no
external API changes.
